[CLEANUP] [CLEAN] Remove src/temporary and relocate placeholder sprite data
- Delete dead src/temporary/map.ts (imported nowhere)
- Move src/temporary/SpriteSheet.ts to @shared/placeholders/spriteSheet.ts
  (live placeholder data used by SpriteProvider) and alias the import
- Remove the now-empty src/temporary directory

Verified: typecheck, lint, and build all pass.
