+++ src/shared/placeholders/spriteSheet.ts
