--- src/temporary/map.ts
@@ -1,35 +1,0 @@
-export const mapData =
-"\
-0404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404041339393939393939393939393939391539393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000040c39393939393939393939393939390c28282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000040d07070739393939393939390707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000040d07070707393939393939070707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000002121000000000000000404000000000000000000000000000004040000000000000000000000000000040d07070707393939393939070707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000210000210000000000000404000000000000002100000000000004040000000021212100000000000000040d07070707393939393939070707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-040000000021000021000000000000040400000000000021210000000000000404000000210000002100000000000004280707070739393939393907070707282727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-040000000021000021000000000000040400000000002100210000000000000404000000000000210000000000000004070707070808080808080808070707072727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000210000210000000000000404000000000000002100000000000004040000000000210000000000000000040c07070707393939393939070707070c2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000002121000000000000000404000000000021212121210000000004040000002121212121000000000000040d07070707393939393939070707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000040d07070707393939393939070707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000040d07070707393939393939070707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000040d07070717393939393939170707070d2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f2727272727272727272727272727272f\
-0400000000000000000000000000000404000000000000000000000000000004040000000000000000000000000000042817171739393939393939391717172828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928\
-040000000000000000000000000000040400000000000000000000000000000404000000000000000000000000000004393939393939393939393939393939391c2a2a2b1c1c2a2a2b1c1c2a2a2b1c1c1c2a2a2b1c1c2a2a2b1c1c2a2a2b1c1c1c2a2a2b3b3b2a2a2b3b3b2a2a2b3b3b3b2a2a2b3b3b2a2a2b3b3b2a2a2b3b3b\
-040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404040404393939393939393939393939393939391c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c\
-1339393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939\
-0c39393939393939393939393939390c28282928282829282828292828282928282829280c282928280c2928282829282828292828282928282829282828292828282928282829280d282928282829282828292828282928282829282828292828280c2828282928282829282828292828282928282829282828292828282928\
-0d07070739393939393939390707070d21212223272121213131313131212727272731210d080b08080d2121272727271a271a27271a2d271c21211a27271a270d212121212131210d2c2d2e272c2d2e2627242521212425212127272425272727270d2121312727272727313127272727272721212727212127272121272727\
-0d07070707393939393939070707070d21213233212131273121212721313127272727270f081b08080f272727272727272727273d271d2727272727272727270d213131272121210d3c3d3e273c3d3e2027343536363435212727273435272727270d2127271a1a272727272727272727272e0c28282928282928280c2c2727\
-0d07070707040404040404070707070d212131302727272727273127272121312727271a1f090909091f1a2727272727272727272727190a0a0a0a0a1c27272728282828272828280d272727272727272027262727272425362026272720272727270d2727271a1a271a27272727272727273e0d3127272727273131283c2727\
-0d070707071c1c1c1c1c1c070707070d21263627272727272727272727273127272727272727272727272727272727272727272727272727272727272727272727272727272727270d272727272727272620263626273435272727272720263627270d271a271a1a271a27270e2727272727270d272727272727272727272727\
-0d07070707393939393939070707072826262027272727272727272721272721272c2d2e2727272727272c1d2e27272727272c2d2e27273d2727273d2727272727272727272727270d27272727272d272727272726272425272727272726272727270d271a271a1a1a1a27270d27272727273d0d27272727272727270c272727\
-0607070707040707070704070707070720242527272716272727272727272127273c3d3e2727270527273c3d3e27272727273c3d3e2727272727272727272727272c2d2e272c2d2e0f2c2d2e27272d272727272736273435272736272726272727270d271a1a1a1a1d2727270d27272727272728272c272a2b272e270d3d2727\
-0c07070707391717171739070707070c303435272727272727272727272731272727272727272727272727272727272727272727272727272727272c2d2e2727273c3d3e273c3d3e1f3c3d3e27272d273636272727272627272720272720272727270d27271d1a1a2d2727270d27272727272727273c273a3b273e270d272727\
-0d07070707393939393939070707070d303026272727272727272727312727312727272727272727272727272727272727272727272727272727273c3d3e272727272727272727272727272727273d272720272727272727272736272720272727270f2727190a0a152727270d27272727272e0c27272727272727270d2c2727\
-0d07070707040404040404070707070d362030212727272727272727272121212727272c2d2e27272c2d2e2727272727273d27273d27273d2727272727272727272727272727272727272727272727272736272727272027272726272727272727271f27273d3d3d3d2727210d21272727273e0d27272727272727270d3c2727\
-0d07070707393939393939070707070d263620202024252627222321213127272727273c3d3e27273c3d3e2727272727210a0a0a0a0a0a0a0a0a272727272727272c2d2e272c2d2e270a0a0a0a0a0a0a272627272727202727273627272727272727272727272727272727210d21272727272728282829282829282828272727\
-0d07070717393939393939170707070d263636202634353636323321212626272727272727272727272727272727272721272727272727272721272727272727273c3d3e273c3d3e2721272721272721272031212131262727272627272727272727272727272727272727210d21272727272721212727212127272121272727\
-2817171739393939393939391717172828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928282829282828292828282928\
-3939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939393939\
-c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1c1\
-";
